create user input element
